refactoring so pre.js doesn't execute both functions
